refactor: add new structure folders in scripts
