+++ index.html
@@ -7,7 +7,7 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <base target="_blank">
   <link rel="stylesheet" href="pages/index.css">
-  <script src="scripts/index.js" type="module"></script>
+  <script src="scripts/pages/index.js" type="module"></script>
   <title>Место</title>
 </head>
 <body class="page">
