+++ scripts/pages/index.js
@@ -1,6 +1,6 @@
-import { dataCards, validationConfig, cardConfig } from './constants.js';
-import Card from './Card.js';
-import FormValidator from './FormValidator.js'
+import { dataCards, validationConfig, cardConfig } from '../utils/constants.js';
+import Card from '../components/Card.js';
+import FormValidator from '../components/FormValidator.js'
 
 
 const KEY_ESC = 'Escape';
